Moved pages into a pages folder
